libnfs.c: fix sign conflict compile warning

+++ libnfs.c
@@ -853,10 +853,10 @@ nfsstat3 nfsio_read(struct nfsio *nfsio, const char *name, char *buf, uint64_t o
 		goto finished;
 	}
 
-	if (offset >= size && size > 0) {
+	if (offset >= (uint64_t) size && (uint64_t) size > 0) {
 		offset = offset % size;
  	}
-	if (offset+len >= size) {
+	if (offset+len >= (uint64_t) size) {
 		offset = 0;
 	}
